Allow keeping data with --keepdata and/or keeping logs with --keeplogs

+++ docker/cleanup.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
-docker volume rm docker_quickstart_data
+if [[ "${@/keepdata}" == "$@" ]]; then
+  docker volume rm docker_quickstart_data
+fi
 docker image prune -f
 docker system prune -f
-sudo find logs -type f -name '*log*' -delete
+if [[ "${@/keeplog}" == "$@" ]]; then
+  sudo find logs -type f -name '*log*' -delete
+fi
